rustapi_module: move issue_219 regression test
