+++ examples/rustapi_module/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod buf_and_str;
 pub mod datetime;
 pub mod dict_iter;
+pub mod misc;
 pub mod objstore;
 pub mod othermod;
 pub mod pyclass_iter;
