+++ examples/rustapi_module/tests/test_misc.py
@@ -1,0 +1,6 @@
+import rustapi_module.misc
+
+
+def test_issue_219():
+    # Should not deadlock
+    rustapi_module.misc.issue_219()
